bug:  Weird sound output on channels 9 to 12 #44
Bug fixed. uint16_t value was going into uint8_t variable.

+++ VoiceNavi/src/play_modes.c
@@ -32,7 +32,7 @@
 #define STOP 0x01
 #define STB 0x02
 
-static uint8_t gpio_prev;
+static uint16_t gpio_prev;
 
 static uint8_t prev_sw = 255;
 static uint8_t irqTriggered;
